+++ ginkgo/core/session/driver.class.php
@@ -10,7 +10,9 @@ use ginkgo\Func;
 use ginkgo\Config;
 
 // 不能非法包含或直接执行
-defined('IN_GINKGO') or exit('Access denied');
+if (!defined('IN_GINKGO')) {
+    return 'Access denied';
+}
 
 // 数据库类型的会话驱动
 abstract class Driver {
@@ -53,7 +55,7 @@ abstract class Driver {
     }
 
 
-    // 配置 since 0.1.4
+    // 配置 since 0.2.0
     public function config($config = array()) {
         $_arr_config   = Config::get('session'); // 取得配置